[a11y] Magnifier
Adds screen magnification.
 * 3x1-tap or 2x3-tap to activate
 * pinch to adjust
 * drag to pan
 * zooms into a region of the screen
 * pixelates
 * animation transitions

Gesture details:

Magnification is enabled by a triple 1-finger tap or a double 3-finger tap.
Once it is enabled, zoom can be adjusted by pinching, and the view can be
dragged to pan (with at least two fingers to start, after which a single
finger will do).

Alternately, magnification can be temporary if the last tap is held down, in
which case panning focuses on the area of the display that would be under the
finger in an unmagnified view.

Test: fx run-test a11y_lib_tests -t magnifier_tests
      on rotated display
Change-Id: I75d2f346357f79b760bfa1e45f0ad2dcdbcba46d