[line input] Factor out abstract base.
Factors out the LineInput API as an abstract base class. Renames the old
LineInputBase to be LineInputEditor since it implements the actual
editing functionality.

This change is to support a future class that will implement this API
to switch between a normal and modal line input.

There should be no behavior change.

Change-Id: I6d2e24a76ecb867dabd411c90704ccd42544b255