[libinet] add methods for mapped V4 addresses
This CL adds three methods to the IpAddress class, one to determine
if a V6 address is a mapped V4 address, another to produce the V4
address from a mapped V4 address, and another to produce a V6 address
by mapping a V4 address. The methods are:

    IpAddress::is_mapped_from_v4()
    IpAddress::mapped_v4_address()
    IpAddress::mapped_as_v6()

V6 addresses mapped from V4 address consist of 5 words of all zeros,
1 word of all ones, and two words that contain the V4 address. That
is "0::ffff:xxxx:xxxx" where the x's make up the V4 address.

TEST: unit test added to verify the new methods
Change-Id: I3f20f7b2ccc0fa6698f18f270381a6779382b131