[cmc] Adds support for cml rights being translated into cm and fidl.
This patchset enables translation and validation from cml (JSON5) into
cm for directory rights. It also enables translation of cm directory
rights into fidl and supports basic validation.

This change is largely mechanical requiring changes to a lot of test
code and updates across the stack to support the notion of rights. One
minor design consideration is I have had to disable part of fuchsia.io2
from building as part of the host toolchain. This is because the host
implementation of the rust bindings do not support stubs for "DebugLog"
along with various other types.

Bug: 36541
Change-Id: Ib904903f8fb8a797a2ef684b988db0737745a74d