[appmgr] Use BufferedPseudoFile instead of UnbufferedPseudoFile for Hub
Currently the Hub uses UnbufferedPseudoFile for URLs and Arguments.
BufferedPseudoFile truncates content to fit within a single I/O
transaction. This transaction appears to be limited to 32 bytes.
This results in the hub providing incomplete information
which impacts programs like 'cs' which scrape the hub.

Switch from UnbufferedPseudoFile to BufferedPseudoFile

Test: Ran under qemu and confirmed output from 'cs' is not cut off.
Change-Id: I1d37a5ea24e24d4d0260e3bb2e8ebe217ce45f3e