[fidl] Make hlcpp roundtrip_test green under v1
With some changes to test_util to add alt_type, we are able to keep most
of the tests in roundtrip_test. Three test are disabled:

The SimpleTable.SerializeAndDeserializeWithReserved test is failing with
a transformer assert.

The XUnion.SimpleUnion test is disabled because the encoded bytes would be
different under v1.

The XUnion.SerializeAndDeserializeSimpleUnion test is disabled because
validation would not work with a v1 coding table.

Bug: 41371
Change-Id: Id6be7383f164606d94d070858f5b4cb715643d85