[mdns] don't panic on early responder channel close
The mdns service had an issue in which it would panic if
PublishServiceInstance was called and the supplied responder channel
was closed while instance probing was still underway. This CL fixes
the issue by checking to see if the InstanceResponder is started
before sending a goodbye message.

TEST: The netemul test has been expanded to repro the early responder
shutdown. This reproed the panic before the fix and succeeds after
the fix.

Change-Id: I0ed685eeda503b4523598c1fcff69eff41f6f6f5