[kernel] clear cmdline entropy from zbi (Reland)
This commits relands CL 335794, which was reverted in CL 337809.
The reason for the revert was because the unit tests were crashing on
Astro.

It turns out that after the system has booted, it is an error to call
`platform_get_ramdisk`, as the memory might not be mapped. It's unclear
whether this was the reason for the test crashes, as I could not
reproduce them locally.

This CL removes the unit test changing it for a `ZX_DEBUG_ASSERT` in
`bootsvc` when the code validates the ZBI handle. Now, we are not only
memsetting the cmdline zbis to 0, we are also setting the header type to
`ZBI_TYPE_DISCARD` and disabling the checksum flag. The assert only
checks that there are no headers with type `ZBI_TYPE_CMDLINE`, as they
should have all been processed in the kernel.

Original change's description:

> [kernel] clear cmdline entropy from zbi.
>
> This CL modifies the zbi parsing process by deleting the cmdline
> arguments after they are processed. One of the arguments might include
> entropy for the kernel cprng and we don't want it to be available to
> other processes with access to the ZBI.
>
> It looks like no other programs are using the CMDLINE zbi objects,
> so this change should be safe. I also used a modified version of
change
> 333450 to scan the memory and look for the entropy in any vmo, the
only
> one that showed up was the zbi vmo.
>
> NOTE that this commit clears ALL the zbi objects of type CMDLINE and
not
> just the parts related to the kernel.entropy-mixin, because that might
> appear in the middle of the payload, and appear multiple times with
> different lengths. Instead, it's easier to just clear everything if it
> is no longer needed.
>
> The whole process is less than ideal, because the ZBI shouldn't be
used
> to pass secrets around, and there are no guarantees about what other
> copies are lying around in memory. Sadly, this is the best we can do
to
> guarantee some entropy during early boot.
>
> I added a test to the global_prng to make sure that the zbi does not
> contain the entropy-mixin parameter. To run it, you can use: `k ut
> global_prng`.
>
> TEST=Tested in Astro, VIM2 and QEMU.
>
> Change-Id: Iba72eaa90740d42d0309735dd4c6925ff7f28755

TEST=Tested in Astro, VIM2 and QEMU.

Change-Id: Ifca97e4309ce6fea9de183185a578b8d3102e73d