[escher] Use precompiled shaders at runtime
Create a new compiler flag on escher called
'escher_use_runtime_glsl', default set to false, which
determines whether or not escher should build with
the glslang and shaderc libraries. If false, escher
will now read in pre-compiled spirv shaders instead
of compiling shaders directly from source-code at runtime.
To accomplish this, the Escher library is now templated, and
has two instantiations available.

The benefit of doing this is to save space on the
Scenic binary, which is nearly halved (17M to 9.6M on
uncompressed builds and 2.18MB to 0.923MB on compressed
builds) when it no longer has to rely on glslang and shaderc.

Testing: New unit tests in shader_program_unittest.cc.
Also, all existing shader program, pipeline creation and
pose-buffer tests continue to work even when built to
use precompiled shaders.

fxb/336637 #done

Change-Id: I6661f5249edf88ad7b1a9e6336bfbf56917f3d2e