[kernel][crypto] Remove dependency on cryptolib.
This CL changes global_prng so that it uses boringssl instead of
cryptolib.

The only usage of cryptolib inside the global_cprng was to calculate a
sha256 hash in `IntegrateCmdlineEntropy`.

To be able to use the SHA256 function in the global_prng, I had to add
the `mem.c` file to the boringssl things that get compiled in kernel,
becase SHA256 calls `OPENSSL_cleanse` to clear its own context .

Change-Id: Ib0c8812e7bd59f66bc9f4f77db12842a098e3222