[build] Allow "disabled" in tags for zbi_test()
This provides a way to mark a ZBI test as disabled so that bots won't
run it, but it's still produced as part of the build so that it can be
run manually.  This is probably useful in general, but the immediate
motivation is to work around the current lack of Infra support for
running ZBI tests on hardware bots and for indicating compatible
hardware environments for the test.

Bug: 10196
Change-Id: I404ccae390093249923c6168ba8fd56a3af7c614