[unification] Produce manifests for binaries in the GN build.
This is the pattern that can be used to build ZN binaries in the GN build.
They are declared as other GN entities but augmented with a local_manifest target
which produces a manifest specific to the binary, to be later included in an
aggregate manifest fed to system assembly.

The pattern is borrowed from the ZN build and can be expanded on in the future to
modernize how we do system assembly in the GN build (though at that point the
intermediate manifest target likely won't be necessary).

Bug: 41352
Change-Id: I8b297c13dcdaaace6cf8fb32a4a75965421db9f1