[fidl][rust] Fix error result representation of Ok(()) in v1
This CL fixes a bug in Rust's representation of error result types in
the v1 wire format (where results are xunions). It now correctly
represents the () success variant with an empty struct, which is a
single 0 byte (padded to 8 bytes in an out-of-line object).

To understand the bug, consider the following two FIDL methods:

    Method() -> ();
    MethodWithError() -> () error int32;

The response of Method is an empty payload (nothing after the header),
whereas the response of MethodWithError is a union (xunion in v1) with
two variants: an empty struct and an int32. The bug was caused by Rust
representing these as () and Result<(), int32> respectively, but only
implementing the empty-payload semantics on ().

This CL fixes the bug by checking for an inline_size of 0 in the
std::result::Result encoding/decoding routines, and instead
encoding/decoding a 0 byte in that case. This works because () is the
only type whose inline_size is 0. It would be cleaner to specialize the
implementations on Result<(), E>, but specialization is not yet
available in stable Rust.

Test: fx run-test rust_fidl_tests

Change-Id: Ic55c144841fea0524e529f576dc771b693ca88f6