[fidl_codec] check string size
Checks that it remains enough data to decode a string. This avoids
allocated very big strings if the message has been corrupted.

For wire_parser_test, only display errors when decoding messages if
we don't expect errors.
We have tests which expect to generate errors and the output was
becomming messy. Now these tests don't display the expected errors.

Bug: 40205

Change-Id: I694da535543ba4e12beed1eb22b54869fa7415d3