[rust][LSC] Fix test failures when compiled with panic=abort
A common theme here is that Rust's test framework spawns a thread for
each test, with a default stack size of 2 MB. Some tests relied on this,
and broke once we were spawning a process for every test (with a default
stack size of 256 kB).

This prepares the tree for panic=abort support, but doesn't add the
option yet.

Change-Id: Ib643b9f778fb1addc6757d478e0491d79ed561bd