[debugger] Plumb up the range setup to x64 helpers.
This basically moves up the logic of ranges to the place where install
watchpoint calls it. This also exposes the information on what range was
actually installed and in which slot.

The watchpoints will use this information to determine whether the
exception has hit due to them and whether they should ignore it because
they're using an overextended range to cover the actual desired range.

This also fixes some un-covered cases in AlignRanges.

TEST=Extended unit test.
     fx run-test debug_agent_tests

Change-Id: Ic500cf8b5639d92477c4787806a6be3ec4611aa1