[pkg-resolver-tests] Fix test to not hang for 30s
The test that verifies packages can still be resolved if a repository
was synced recently was implemented by stopping the repository http
server, but that results in the package resolve operation waiting for
the 30 second connect timeout in amber.  This change modifies the test
to instead re-configure the http server to fail all requests so the test
does not need to wait on a timeout.

Change-Id: I696a1a19c1450062fcf120d6a43fb5522d7fbd4e