[debugger] Watchpoints track actual installed range.
The architecture might decide to use a bigger range for the actual
watchpoint installation (eg. to cover for unalinged ranges by issuing a
bigger aligned one).

This means that the watchpoint needs to know the range it is supposed to
watch and the actual range it is watching, in order to correctly
interpret the exceptions.

TEST=Added unit test.
     fx run-test debug_agent_tests

Change-Id: Icc2826d85ab32851a067fdd3d1d33040d2a4a817