[debugger] Skip compiler code after "finish"
Thread controllers generally need to skip over code marked with "line 0"
because it's generated by the compiler. If execution stops on one of
these lines, the user will see assembly code which is normally
undesirable.

The "finish" thread controller can land on such code and leave the user
at an unsymbolized address. This detects "line 0" code immediately after
a "finish" command and steps over it.

Clarifies "line 0" handling in some comments. Updates the module symbols
to avoid setting a file name for "line 0" entries because the file name
is irrelevant and misleading.

BUG=41427

Change-Id: I5e121920886301bc5d88501a0e2be9989519997c