[zircon] Gate ktrace_init() on kernel.enable-debugging-syscalls
Make ktrace_init() early-out when kernel.enable-debugging-syscalls is
false. In that state, the syscalls used to access ktrace data are
already disabled, so there is no point to reserving the ktrace buffer
which defaults to 32M.

Add integration tests to confirm that uapp/ktrace does/doesn't work when
this kernel command line argument is in effect.

Bug: 41380
Change-Id: I7ae597406dc17edf7f60535c2a4e1e9be84d2614