[PowerManager] Implement ThermalPolicy node
Fill out the ThermalPolicy node which currently exists as a placeholder.
The node communicates with the TemperatureHandler and
CpuStatsHandler nodes to gather the data required for the thermal policy
calculations, and with the CpuControlHandler to impose CPU performance
limits for thermal mitigation. The thermal policy operates as a PID
control loop, where temperature is the measured process variable and
available power is the manipulated control variable.

Bug: 41254
Test: fx run-test power_manager_tests
Change-Id: Ic00556e438ec079a6a41685b864996636d997ad9