[pkg-resolver-tests] Decrypt repo template keys
The checked-in template repository for use in the package resolver
integration tests stored its private role keys in encrypted form, which
slow down the tests without much benefit.  This change regenerates the
template repository with decrypted keys, reducing the runtime of the
pkg-resolver-tests on a NUC from 14 seconds to 10 seconds.

Test: pkg-resolver-tests
Change-Id: I8a671206c624b4bf3576285e8e40d28bcc794635