[fidl][transformer] New day, new fixes
(1) When we transform a table, we may encounter unknown fields, i.e.
fields which are present in the table at a given ordinal, but for which
ordinal we do not have any details. These unknown members must be passed
as such to `TransformEnvelope`.

(2) When tranforming unknown envelopes, we must both copy the envelope
header and the envelope data.

(3) It's also important to properly report up, using the
`TraversalResult`, both the out-of-line bytes transformed, and the
number of handles transformed.

Bug: 41371
Test: added for 1, 2, and 3
Test: fx build zircon && fx run-host-tests fidl-test
Change-Id: I10850ae5fec21171dff68360c90b81f68d951c09