[fidl][transformer] Refactoring debug handling
The impetus for this refactoring is that we were not setting the type
under some condition, leading to a segfault when formatting said type.
Here, we're extending the lifecycle of the `DebugInfo` object to be
scoped to `fidl_transform`, and can therefore leverage this object to be
a helper as we gather debug information. This simplifies the code, and
makes it safer.

Added a test to ensure that we fail and emit the expected error message
if we attempt to transform with an unknown transformation. Prior to this
change, this would segfault.

Test: fx build zircon && fx run-host-tests fidl-test
Change-Id: I4b5871328bfd80b84def6f8ae997e4f2e0d7cd44