[ledger] Add flags for diff support level in test cloud provider
The goal of this change is to be able to simulate in integration tests:
 - a group of Ledger instances without compatibility support talking to
 each other sucessfully. This is already testable with the current
 behavior of the test cloud provider and the Ledger flag introduced in
 change Ia5387d0f7415abc98336b3fc0273d4b8d40d207a.
 - a group of Ledger instances, some of them supporting diffs and some
 of them not supporting it (or the cloud provider starting to support
 them), with compatibility support. A superset of these situations can
 be simulated by randomly refusing to accept diffs sent by clients in
 the cloud provider.
This change introduces a flag controllable through FIDL to enable the
latter behavior in the in-memory cloud provider.
To get reproducibility in integration tests, the C++ library interface
now requires a random seed to create a cloud factory.
A subsequent change will make integration tests run in these two situations.

Test: crate::session::diff_mode fakes a random number generator
Change-Id: I126e600192dd2e7103108a37444c3eccd1309764