[debugger] Step into inlines properly.
Previously inline stepping worked the way one would expect only if the
inline function was the first thing on the line the user executes "step"
from.

If the user started on some code on the same line as the inline call, it
would stop immediately before the inline call. "Step" would step into an
inline call only if it was the first instruction on the current line.

This change makes "step" step into the inline if the inline's call
location is the same file/line as the line being stepped. If it's a
different file/line, the stack will look like it's immediately before
the call.

BUG=41344

Change-Id: I5516a1598be09c9e1e4709766734071d16813a84