[tools][dev_finder] Adds netboot support.
This adds netboot to the protocols dev_finder uses for device discovery.

This also lowers the timeout default to 200ms as 2000ms is quite a long
timeout.

Inside netboot.go there is also a change to prevent unnecessary looping
after a connection is closed (this also gets rid of spurious error logs)
when running discovery so that, in the event of parsing errors, the
client doesn't need to ignore Target.Error when it is raised.

Bug: 37637
Tests: fx set core.x64 --with-base //tools/:tests && fx build && \
      $FUCHSIA_DIR/out/default/host_x64/dev_finder_tests
Change-Id: I752b0d4c50a51b867798a5b381dc3319cbf48394