[debugger] Don't skip prologues for inline frames.
The change that automatically skipped function prologues introduced a
regression by trying to skip prologues for inline functions (which of
course have no prologue) as well as physical functions (which do). The
result is that the first statement of every inline function would be
magically skipped.

This change stops skipping prologues for inlines.

Change-Id: I79a10df63bfee18d8cbcf334056575541ff344ae