[componentmgr] Simplify hooks
All hooks use a single callback at this point so there's no point in
having a different callback for each event type.

This CL replaces HookRegistration which takes a single event type
and a callback for it with HooksRegistration which takes multiple event
types for a single callback.

This CL also makes it clear that hooks take in a weak reference and not
a strong one to clarify the contract with Hooks::install.

Change-Id: Id1a4ae1742d236b736a605e85ea8139af92c0019