[unification] Remove the ability to reference //zircon/system/ulib in the GN build.
While this scheme works, in practice it is not actually useful as shared
libraries have to be singled out. Indeed, there can only be one version of
a given shared library used throughout the build because there can only be
one version available at runtime (at least for /boot, which is where a
majority of the ZN-built shared libraries end up among other places). Even
if the two builds were to yield the exact same binary, the different paths
on the build host would likely throw off the system assembly process.

And when a given shared library is only referenced in the GN build, it is
just as easy to migrate that library's code out of the ZN build without
resorting to a springboard under//zircon/public/lib.

Bug: 35659
Change-Id: I96d3ed3514d4cf6b12fba60fbcb7bf217b6a1819