[pseudo-fs-mt] mut_pseudo_directory! macro
Now that we have to kind of "simple" directories, it is convenient to
have two helper macros to construct them.  An alternative would be to
intorduce a parameter to be passed to `pseudo_directory!`.  But I
figured, at least while there are only two possibilities, a different
name seems to be a bit better - it makes it more explicit that the
constructed object is different.

Change-Id: I603c0c0e5d6cd762be0cedf0539d22607e5b4a46