[audio_core] Output mixer uses input streams directly.
This removes the AudioLinkRingBufferSource and AudioLinkPacketSource
classes. Links now optionally have a Stream associated with them, with
the only current implementation of Stream being the PacketQueue.

This implementation doesn't require the downcasts previously used,
since the AudioRendererImpl simply retains a reference to the concrete
PacketQueue type while the generic Stream interface is exposed by the
AudioLink.

Test: fx run-test audio_core_unittests
      fx run-test audio_pipeline_tests
Change-Id: I6c918e343729e789444afc82f0b19cffd526e543