[inspect][rust] Fix bug in Inspect reading code that wasn't reading lazy nodes within children in lazy nodes.
The test case `reader/test_reader.rs:tests:test_read` illustrates this case.
The core of the fix is in `reader/test_reader.rs:39` that expands
the links that the children of the snapshot node have.
Introduced the ReadableTree trait to allow the reader code to work for
both Inspector and TreeProxy. This CL only includes the Inspector
implementation. The TreeProxy one comes in the follow up.
Also refactored the code to be a more concise and clear algorithm, except that it's recursive now.

TESTED=fx run-test fuchsia_inspect_tests

Change-Id: I57eb1fef4aece84ab462ecc9f420c5d77ab6352e