[ledger] GC Policies: Add policy to schedule GC only on tree node roots.
On the eager GC policy, garbage collection was scheduled when an object had 0 live
references. However this leads to many non-succeeding deletions.

After this CL, GC scheduling is triggered after a commit has been pruned or marked
as synced to the cloud and only for the root tree node: if that one is successfully
deleted, the scheduling algorithm will recursively find the rest of the objectes
referenced and try to delete them.

After a commit is pruned or marked as synced, GC doesn't start directly: we are
also waiting until all its live reference count reaches 0.

Note however, that this approach will not immediatly garbage-collect objects that
have been added in storage, but are not part of a commit: there will need to be a
general GC executed periodically in the background to find leftover objects.

Change-Id: I7baba5ec39b638b921d28619b7cb5472ada464d2