[audio_core] Optimize output-to-uint8
This CL rearranges the order of calculations used when converting from
our internal representation (float) to an output format of uint8_t.
(The uint8 format is rarely-used either as a driver format or a client
format, but it is still occasionally seen in certain older WAV files.)

Specifically, this change pulls an addition out of one clamp subclause
and instead applies the addition to the result of that clamp. This
leads to two benefits: all three operands of the clamp are now signed,
and the addition is now performed on entities of identical type.

With zero functional change, this improves the OutputProducer's speed
(when converting to uint8) by 20-25%, bringing this format's
performance closer to (but still not equal to) that of the int16,
int24-in-32 and float output formats.

Test: no behavior change, covered by the significant number of
existing tests in audio_mixer_unittests.

Change-Id: I8a7a45b71171ad025aa1f1bf57440745299594e8