[camera] Add camera stream constraints
Producing configurations for camera HALs is difficult
because of the deeply nested fidl data types involved.
However, the amount of information they store tends to be
pretty small (modulo some default options).

This CL introduces a class that generates Camera Configs
from a smaller data structure.  Currently this is used
to simplify camera manager test code, but it may have applications
elsewhere in the camera stack.

Test: This is part of the code used to test the camera manager.
It is incorporated into the video-device-client-test in a later
CL.

Change-Id: I82858132b701df675464be78f33d47c16c2ef187