[go][kazoo][syscalls] Add Go VDSO keys and calls outputs
This is the last known user of abigen data. go_vdso_output.cc is pretty
much a transliteration of
https://fuchsia.googlesource.com/third_party/go/+/refs/heads/master/src/runtime/mkfuchsiavdso.go.

This should allow us to remove the abigen parsing in our third_party/go
fork, and make a script in that repo to update syscalls based in the
source data in //zircon/syscalls.

No intended functionality change; this output won't be used until we
manually roll in the Go repo.

Test: new in go_vdso_output_test.cc
Bug: 33668
Change-Id: I95cdece2851450e3c51caff808dd3f12e5923c26