[nmh][network_manager] Add oir support.
Adds Online Insertion Removal support.

It detects when a new port in added to the system and gives it to
netstack for it to use.
Removal is not yet supported (nothing is done when removed).
It also configures the device as a router, with one port as a WAN
interface and the rest a past of a switched LAN interface.
The configuration is compiled in. This will be replaces by an actual
configuration file that is loaded when the device starts in a comming
CL.

Change-Id: I50af8255e291fac36f0343e190fb8f9a81ea526d