[fdf][coordinator] Do not log error on devhost disconnect if expected.
When the devhost is processing a CompleteRemoval request, it will reply
to the request and then close the connection.

On the devcoordinator side, we will remove the device and mark it as
dead, but we also detect the disconnect and attempt to remove the
device again, which manifests as a log error when we try to disconnect
a USB device.

Change-Id: Idda64b347b278e48fb81f0dcd13f8124b3056d28