[bt][test] Allow Test Harness types to be composed
Change the Test Harness implementation to use separate `init` and
`terminate` methods instead of a combined `run` method. This
decomposition allows multiple harnesses to be composed by piecewise
composition of `init` and `terminate`.

The `init` method returns both the harness, and also two secondary
pieces of data:
* an `Env` that contains required environmental resources that should be
persisted until `terminate` is called, but that should be hidden from
the test
* A `Runner` future that represents asynchronous work that must be
executed in order for the harness to function correctly

This allows us to remove uses of fasync::spawn in all our Test
Harnesses. We also pull out the execution (via
`executor::run_singlethreaded`) into a run_test function rather than in
the Harness implementation itself; this allows Harnesses to potentially
be executed in other ways.

An impl of `Harness` is added for a tuple of Harnesses, so that tests
may ask for (and receive) multiple harnesses if so required.

Testability: This is itself testing functionality. The validity of this
refactor is asserted by running the tests that depend on it:

  $ fx run-test bluetooth-tests -t bt-integration-tests

Change-Id: I7ec8fb48a458f0da516ba87bab730a834d697ed2