[debugger] Enhance sym-info command.
Print brief info for more types of objects, including types and
functions. This can be enhanced in the future if we want to know more.

Demangle names. This allows the command to be used as a general
demangler even if no match is found, and will also find names based on
the mangled one.

Change-Id: Ic58f2494d4ed311de4d0073ad5c5ba0086bee209