[debugger] Remove symbol output from variable finding.
EvalContext::GetVariableValue and related functions computed the value
of a variable, and also returned the matched symbol in the callback. The
purpose of this was for the sym-info introspection command.

The sym-info command now uses FindName directly which gives it more
information and means the matched object is not needed on the
GetVariableValue code path any more. The callback argument is now
removed and surrounding code simplified.

Change-Id: I8ca6c3b2d30f3d50060002f6ff534158931a63a0