[bootsvc] decommit the last page of the zbi.
This CL makes sure that the last page of the zbi is decommitted, if it's
not needed anymore.

The previous implementation had an issue that if the last element was
not aligned to the zbi size, the last page was not decommitted because
`DiscardItem` rounded down discard_end to `PAGE_SIZE`.

This change rounds up discard_end before calling DiscardItem so that the
last page is always considered.

The code scans the zbi vmo looking at one element at a time. If an
element can be discarded, discard_end gets advanced to the end of the
element. If an element should not be discarded, everything from
`discard_begin` to `discard_end` gets discarded, and `discard_begin`\
points to the end of the element.

At the end of the loop, `discard_end` is either pointing to the end
of the last element of the zbi, or it is pointing to the beginning of
the last saved element. In the latter case, it should be true that
`discard_end < discard_begin`, and that means that there's nothing to
discard. Otherwise, are in the first case, meaning that we reached the
last zbi element and it should be discarded, and it is safe to discard
up until the end of the vmo.

Change-Id: I3edfec43bad7a9eacc087a046ed1161eb3b25c8c