[audio_core] Optimize SincSampler by using shorter ChannelStrip
Using shorter internal buffers makes the SincSampler about 15% faster
in its microbenchmarks (the '--profile' flag). This is presumably
because of cache/locality effects, and slightly smaller memory
footprint. Specifically, the ChannelStrip used by SincSampler has been
shortened from 1000 frames to 680 frames. Various lengths were
profiled (280, 560, 640, 680, 700, 720, 1000, 1344), with this one
performing best.

Note: this value directly sets the maximum allowable downsampling
rate; specifically that ratio is precisely length/28 (so for a length
of 280, the maximum downsampling ratio would be 10:1). With a value of
680, greater than 24:1 downsampling is possible. Thus even if input
hardware runs at 192 kHz, we would still be able to downsample to less
than 8 kHz.

Test: extensive test suites still pass
    fx run-test audio_fidelity_tests -- --profile --full
    fx run-test audio_mixer_unittests
    fx run-test audio_core_unittests

Change-Id: Ide9abad5aaf4c298dd0ecf4c68e23b8d2992f69f