[PowerManager] Use unique TemperatureHandler instance per driver
Currently, a single TemperatureHandler node instance is the sole entity
to interact with any temperature sensor in the system. The instance
stores the open channels to each unique driver in a hashmap. In this
arrangement, if a different node is requesting temperature from the
TemperatureHandler node it also needs to specify the driver path it's
interested in. I think a better approach is for there to be a separate
TemperatureHandler instance for each temperature driver, and each
instance only interfaces with one driver. This approach lets us simplify
the TemperatureHandler code, especially as this node grows in complexity
to support additional functionality such as multiple client nodes asking
for temperature from a single driver.

Bug: 41385
Test: fx run-test power_manager_tests
Change-Id: I9551370950dffa0d38cbe611f89677f06b788e28