[audio_core] Move renderer timeline function into Stream.
This moves the 'reference clock to fractional frame' conversion
function into the Stream interface so that the mixer does not need to
query the renderer object directly for this value.

Since one function is shared by all streams attached to a renderer, we
use a ref-counted wrapper around the timeline function that allows the
renderer to update the function as needed (ex: during pause).

Test: fx run-test audio_core_unittests
      fx run-test audio_pipeline_tests
Change-Id: I1c05eefa8bd586d2882b52733e736b41cc15ae0e