[gidl] Refactor GIDL IR and generate more transformer tests
This CL makes the transformer backend of GIDL generate additional tests:

* For GIDL decode success tests that only provide one set of bytes,
  generate old->v1 and v1->old success cases (they should be no-ops).

* For GIDL decode failure tests, generate old->v1 and v1->old failure
  cases ensuring that fidl_transform does not crash.

* Generate tests from all .gidl files, not just transformer.gidl.

To make this work, this CL adds v1 bytes to all existing tests that use
unions. Now, all tests either specify both old and v1 bytes, or only
specify old bytes because v1 is the same. We plan to improve the syntax
in the future to make this explicit.

This CL also refactors GIDL's intermediate structures to simplify the
transformer test generation. Instead of having multiple objects with the
same Name but differing WireFormat and Bytes, we now use a single object
with an []Encoding field, where Encoding is a struct bundling together
WireFormat and []byte. We use this instead of a map[WireFormat][]byte to
ensure deterministic test ordering.

Test: regen and run all GIDL tests
Test: fx run-host-tests gidl_parser_test

Fixed: 39579 (EncodeFailure for all wire formats)

Change-Id: I56e4a10e8bdb3fae75419dda063663d1e992572f