[audio_core] Remove superfluous heap allocations.
Many places we were depending on AudioStreamTypePtr instead of
AudioStreamType. The *Ptr variant is simply an alias for the
unique_ptr<> wrapped version of the structure, but most consumers don't
require heap allocation for any reason.

Updates functions that operate on AudioStreamTypes to take a
'const AudioStreamType&` and also replace some AudioStreamTypePtrs with
in-line AudioStreamTypes

Test: fx run-test audio_core_unittests
Change-Id: I55c28cd3121b0bcbfba53c328692b526cc16b147