[fidl_codec] display partially decoded
Now, we can display partially decoded messages (including messages which
failed to decoded).

So, when we are not able to decode a request or a response, we now
display what we have been able to decode for the request and the
response (on top of the errors we had while decoding).

This way, it's easier to find what was not decoded correcly.

Bug: 40205

Change-Id: I373978cc801446989607a19535f3cb3a1679f67f