[tools][net][netboot] Fix test flake.
This fixes a test log flake that happens when a connection is closed
(occasionally the fake server loop still attempts to read on its port,
causing a finished test to create a test log after a test is completed).

This error should not be fatal, and is expected to happen on occasion.

Bug: 40985
Change-Id: I1215120f76462c550c5f5934ffdc115477945f67