[camera] Add VideoDeviceClient
The VideoDeviceClient class keeps track of device
properties and configurations for the camera manager.
This CL just adds the class and the unit tests, but does
not incorporate the VideoDeviceClient into the camera manager.

This CL was previously: Ia0d4f85a2859a9864e976fb1d44917937260b261
which was reverted: I89fb8f1c7801b7aeba2a20a6ee870e9f00a1f1f6
Bug: 41675

This CL contains a fix:  When creating local variables with bindings,
the async loop must be shut down in the test to prevent a race where
both bindings are destructing and messaging each other.
I was able to reproduce the flake in the asan variant, and verify that
it does not re-occur.

Testing: A new test is added: video-device-client-test.

MULTIPLY: `[
  {
    "target": "//src/camera:camera_tests",
    "os": "fuchsia",
    "total_runs": 1000
  }
]`

Change-Id: If237c6f26ebaf4e450fa56cb9c52ddb9da67f35e