[fshost] swap order of inheritance for registry vnode
there is a bug in llcpp where it does some kind of reinterpret_cast that
assumes the pointer is stable, but it's actually not stable moving
around the class hierarchy like that, because it depends on inheritance
order. or something like that. swapping them makes the service actually
work.

a test is provided that sets up the registry vnode as a fidl service and
attempts to call it. it fails when the interfaces are not swapped.

there's also some minor changes fx format-code threw in there as well.

Test: /system/test/ddk/fshost-test-test
Change-Id: I5d20d4081ee713ed0d931c340cd3eb8ec3c0677d