[audio_core] Rewrite audio_device_tests
audio_device_tests depedended on the concept of a default device to
detect device additions and removals.

We are removing this behavior, and writing out of proc tests in Rust
moving forward.

This change replaces tests that did not cover default device order. It
also implements the hermetic audio test framework in Rust, which will
be promoted to a lib soon for more out of proc tests.

MULTIPLY: `[{
    "target": "//src/media/audio:audio_device_tests",
    "total_runs": 100
}]`

Bug: 41059
Test: fx run-test audio_device_tests
Change-Id: I8b44bbfcada94892cb0a64ac90d81c653e3a125c