[FDF]Change the way resume task works
At the moment, resume task is spawned on the parent, the parent resumes
itself then issues resume taks on all its children and waits for them to
complete. This approach will work well from system resume's perspective.
If we create resume task on the tip of the tree it will propagate to
various branches and all the devices will be resumed.

However, this approach will not work for device resume. For a particular
device to be resumed, we want the resume to be bubbled up to all the
ancestors in the chain.

This change fixes the resume task in a way that it will address both the
cases. For System resume, we issue the resume task on all the devices
and let each one of them bubble up and complete the entire system
resume.

Bug: 38111
Test: runtests -t devcoordinator_test

Change-Id: I0e25b815bf5800515568a62d20ff00d632727b0b