[debugger] Allow better backward symbol references.
Adds a variant of a LazySymbol that doesn't cache an owning reference to
the constructed object. This is used for backward references to the
parent of a symbol object to prevent refcount cycles.

A helper object is added to allow setting the parent in a safer way
using an explicit object for testing purposes.

Change-Id: I64fb2c8dc1883a8214fa5326395c8a21269e7bf7