[camera][calibrate] Factory FIDL interface
This CL adds the first draft for the on-device FIDL interface that is meant to
interact with the SL4F Factory Facade (fxr/331871).

The two important endpoints we want to expose to a host are:
* Instruct the device-under-test to take a picture.
* Instruct the device-under-test to write some tuning calibration data
to its persistent storage.

The rest of the endpoints are for camera detection and setup.

Tested: does not break previously written tests. There is no
implementation yet so there is no new behavior to test.

Change-Id: I2f85a04deeab69980ff39a9149550a02cffd0250