[benchmarks] Add subset of Ledger tests to set of perf tests run via SL4F
This invocation is based on those in
peridot/tests/benchmarks/benchmarks.cc, which runs on the peridot-perf
bot.  This is part of moving perf tests to being run via SL4F instead
of via peridot/tests/benchmarks/benchmarks.cc.

Split out an addTspecTest() function for reusing.

Declare the dependency on the ledger executable.  Add a forwarding
dependency to deal with GN target visibility.  For comparison: for the
old case (the peridot-perf bot), the dependency on //src/ledger/bin is
declared in peridot/packages/BUILD.gn, but that bypasses the
visibility check with a list of exceptions added by build/package.gni.

The full set of Ledger tests goes over the current 5 minute time limit
for a test executable, so until we work out how to deal with that, we
just add a single Ledger test here, so that we can at least land part
of this transition.

Bug: 23091
Test: on CQ: terminal-x64-release trybot;
  local testing: "fx set terminal.x64", "fx run-e2e-tests perf_test"
Change-Id: Id0ed9c81b199d29041d71c5d0f80d3646c9e2854